Revert "Kalyan macos compile"
